YD-107 Two test case corrections: * Don't try to delete a user if creation failed * Test the addUserResponse.status too

+++ appservice/src/intTest/groovy/nu/yona/server/UserTest.groovy
@@ -39,7 +39,9 @@ class UserTest extends Specification {
 			testUser(response.responseData, true)
 
 		cleanup:
-			appService.deleteUser(appService.stripQueryString(response.responseData._links.self.href), password)
+			if (response.status == 201) {
+				appService.deleteUser(appService.stripQueryString(response.responseData._links.self.href), password)
+			}
 	}
 
 	def 'Send mobile number confirmation code'(){
@@ -53,8 +55,9 @@ class UserTest extends Specification {
 
 		then:
 			confirmationCode != null
-			response.status == 200
+			userAddResponse.status == 201
 			userAddResponse.responseData.confirmed == false
+			response.status == 200
 			response.responseData.confirmed == true
 
 		cleanup:
